refactor(server): split test_cfstore into test_config and test_processor
Config tests and processor tests had no reason to share a file. Moves
make_adapter to conftest.py so both test modules can use it without
duplication.
